+++ providers/modules/dzen_ru/provider.py
@@ -13,7 +13,7 @@ from providers.base_module import PlatformModule
 class DzenRuProvider(PlatformModule):
     name = 'dzen.ru'
     status = 'not_configured'
-    host_suffixes = ('dzen.ru',)
-    ie_names = ('dzen.ru', 'dzen.ru:channel')
+    host_suffixes = ('dzen.ru', 'www.dzen.ru', 'zen.yandex.ru')
+    ie_names = ('dzen.ru', 'Dzen', 'ZenYandex', 'dzen.ru:channel')
     source = "catalog"
     description = 'Дзен (dzen) formerly Яндекс.Дзен (Yandex Zen)'